Trying to figure out count of elements
I thought this may be able to count how many titles there are in the xml, but it turns out it counts the number of characters there are instead. I'm trying to figure out how to count the number of titles, artists, prices, etc so that I can make sure it is all equal to the number of CDs there are.